move socket handlers in module
